project/backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==21.2.0
cryptography==42.0.5
